+++ packages/frontend/mobile/src/views/sign-in/layout.css.ts
@@ -1,5 +1,3 @@
-import { cssVar } from '@toeverything/theme';
-import { cssVarV2 } from '@toeverything/theme/v2';
 import { style } from '@vanilla-extract/css';
 
 export const root = style({
@@ -17,45 +15,4 @@ export const content = style({
   flexDirection: 'column',
   alignItems: 'stretch',
   gap: 24,
-});
-
-export const skipDivider = style({
-  display: 'flex',
-  gap: 12,
-  alignItems: 'center',
-  height: 20,
-  marginTop: 12,
-  marginBottom: 12,
-});
-
-export const skipDividerLine = style({
-  flex: 1,
-  height: 0,
-  borderBottom: `1px solid ${cssVarV2('layer/insideBorder/border')}`,
-});
-
-export const skipDividerText = style({
-  color: cssVarV2('text/secondary'),
-  fontSize: cssVar('fontXs'),
-});
-
-export const skipText = style({
-  color: cssVarV2('text/primary'),
-  fontSize: cssVar('fontXs'),
-  fontWeight: 500,
-});
-
-export const skipLink = style({
-  color: cssVarV2('text/link'),
-  fontSize: cssVar('fontXs'),
-});
-
-export const skipLinkIcon = style({
-  color: cssVarV2('text/link'),
-});
-
-export const skipSection = style({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
 });
